Add new test modules for V1, restructure
