+++ plutus-benchmark/script-contexts/test/V3/9.6/checkScriptContext1-4.budget.golden
