+++ plutus-benchmark/script-contexts/test/V3/9.6/checkScriptContext2-4.eval.golden
