+++ plutus-benchmark/script-contexts/test/V3/9.6/checkScriptContext2.size.golden
