+++ plutus-benchmark/script-contexts/test/V3/9.6/checkScriptContextEqualityData-20.budget.golden
