+++ plutus-benchmark/script-contexts/test/V3/Data/9.6/checkScriptContext2.pir.golden
