+++ plutus-benchmark/script-contexts/test/V3/Data/9.6/checkScriptContextEqualityOverhead-20.budget.golden
